+++ src/test/java/de/dm/intellij/liferay/language/properties/LiferayPortalPropertiesCompletionContributorTest.java
@@ -1,0 +1,65 @@
+package de.dm.intellij.liferay.language.properties;
+
+import com.intellij.codeInsight.completion.CompletionType;
+import com.intellij.openapi.module.Module;
+import com.intellij.openapi.projectRoots.Sdk;
+import com.intellij.openapi.roots.ContentEntry;
+import com.intellij.openapi.roots.ModifiableRootModel;
+import com.intellij.openapi.util.Disposer;
+import com.intellij.openapi.vfs.newvfs.impl.VfsRootAccess;
+import com.intellij.testFramework.IdeaTestUtil;
+import com.intellij.testFramework.LightProjectDescriptor;
+import com.intellij.testFramework.fixtures.BasePlatformTestCase;
+import com.intellij.testFramework.fixtures.DefaultLightProjectDescriptor;
+import com.intellij.util.PathUtil;
+import de.dm.intellij.liferay.language.freemarker.LiferayFtlVariableProviderTest;
+import de.dm.intellij.liferay.module.LiferayModuleComponent;
+import org.jetbrains.annotations.NotNull;
+
+import java.io.File;
+import java.net.URL;
+import java.util.List;
+
+public class LiferayPortalPropertiesCompletionContributorTest extends BasePlatformTestCase {
+
+    private static final LightProjectDescriptor MY_PROJECT_DESCRIPTOR = new DefaultLightProjectDescriptor() {
+
+        @Override
+        public void configureModule(@NotNull Module module, @NotNull ModifiableRootModel model, @NotNull ContentEntry contentEntry) {
+            LiferayModuleComponent.getInstance(module).setLiferayVersion("7.0.6");
+
+            URL resource = LiferayFtlVariableProviderTest.class.getResource("/com/liferay/properties/70/portal.properties");
+            String resourcePath = PathUtil.toSystemIndependentName(new File(resource.getFile()).getAbsolutePath());
+            VfsRootAccess.allowRootAccess(Disposer.newDisposable(), resourcePath );
+        }
+
+        @Override
+        public Sdk getSdk() {
+            return IdeaTestUtil.getMockJdk18();
+        }
+    };
+
+    @Override
+    protected void setUp() throws Exception {
+        super.setUp();
+    }
+
+    @NotNull
+    @Override
+    protected LightProjectDescriptor getProjectDescriptor() {
+        return MY_PROJECT_DESCRIPTOR;
+    }
+
+    @Override
+    protected String getTestDataPath() {
+        return "testdata/de/dm/intellij/liferay/language/properties/LiferayPortalPropertiesCompletionContributorTest";
+    }
+
+    public void testPortalPropertiesCodeCompletion() {
+        myFixture.configureByFiles("portal-ext.properties");
+        myFixture.complete(CompletionType.BASIC, 1);
+        List<String> strings = myFixture.getLookupElementStrings();
+        assertTrue(strings.contains("admin.email.from.name"));
+    }
+
+}